UBUNTU: SAUCE: (no-up) apparmor: fix apparmor spams log with warning message
This is a fix to code that is not upstream.

Remove label check warning that is enforcing a condition that is not
yet always valid in the trusty version of apparmor. The check leaked
in from later patches in the -dev tree and does not belong in the
trusty code base.

BugLink: http://bugs.launchpad.net/bugs/1308761

Signed-off-by: John Johansen <john.johansen@canonical.com>